refactor: AddChatLayout 컴포넌트 리팩터링
