+++ src/app/(providers)/(root)/[workspaceId]/chat/add/_components/AddChatLayout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AddChatLayout';
