+++ src/app/(providers)/(root)/[workspaceId]/chat/add/layout.tsx
@@ -3,8 +3,8 @@
 import type { StrictPropsWithChildren } from '@/types/common';
 import { SearchUsersProvider } from './_provider/SearchUsersProvider';
 
-const AddChatLayout = ({ children }: StrictPropsWithChildren) => {
+const AddChatProviderLayout = ({ children }: StrictPropsWithChildren) => {
   return <SearchUsersProvider>{children}</SearchUsersProvider>;
 };
 
-export default AddChatLayout;
+export default AddChatProviderLayout;
